% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_detector.R
\name{gap_detector}
\alias{gap_detector}
\title{Detecta periodos de sell out nao observado de um determinado produto. necessário construir a lista de cnpjs a analisar, uma tabela chamada 'aux' contendo campos cnpj,dtt apenas das vendas do item de interesse e uma tabela chamada dt2_data com todas as datas presentes na base original em ordem crescente.}
\usage{
gap_detector(cli_nome)
}
\arguments{
\item{cli_nome}{cnpj que sera analisado}
}
\value{
dataframe com periodos de ruptura do item ou conjunto de item selecionado na construção do dataframe 'aux'. Onde prim é a primeira semana de ruptura, lastbut1 é a última semana de ruptura e gapsize é o duração da ruptura em dias
}
\description{
Detecta periodos de sell out nao observado de um determinado produto. necessário construir a lista de cnpjs a analisar, uma tabela chamada 'aux' contendo campos cnpj,dtt apenas das vendas do item de interesse e uma tabela chamada dt2_data com todas as datas presentes na base original em ordem crescente.
}
