% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_funcao.R
\name{switch_funcao}
\alias{switch_funcao}
\title{funcao facilitadora para gerar a migracao}
\usage{
switch_funcao(
  antes_base,
  depois_base,
  CNPJ_input = "cnpj",
  PROD_ID_input = "sku_desc",
  prim_input = NA
)
}
\description{
funcao facilitadora para gerar a migracao
}
